### 2022-11-18 #4854 @jbytecode
👋👋👋 Dear @msalibian and @mingzehuang    👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**robnptests -- An R package for robust two-sample location and variability tests**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4854).

Thank you in advance!
