### 2022-11-18 #4947 @editorialbot
[REVIEW]: robnptests -- An R package for robust two-sample location and variability tests
**Submitting author:** @s-abbas (<a href="http://orcid.org/0000-0001-9162-9792">Sermad Abbas</a>)
**Repository:** https://github.com/s-abbas/robnptests
**Branch with paper.md** (empty if default branch): develop
**Version:** v1.1.0
