### 2022-11-29 #4947 @msalibian
I have finished my revision. This is a nice contribution, and my suggestions are rather minor, some of them perhaps just due to my not understanding the requirements on the checklist.

Regarding the checklist:
- The LICENSE file is missing.
- I did not find Community guidelines (for reporting issues or contributing to the project) in the documentation (they are there on the GitHub repo). This is not typically present in R packages stored on CRAN, so I'm inclined to not require it here.
- Automated tests to verify functionality, I'm happy with the "seeded" examples in one vignette, even if these are not proper unit tests, they do provide a way to check the behaviour of the software.
- Quality of writing: the paper is well written, my suggestions here are fairly minor and mostly about disambiguation (see below).

Specific suggestions for the text of the paper:
- Title: "location and variability tests", I suggest using "location and dispersion tests". Variability may be perceived as implying the existence of second moments. In the same spirit as the authors use "location", I'd suggest using "dispersion" instead of variability.
- Summary: "high efficiency", efficiency is a concept usually associated with point estimators, not tests. Can this be rephrased in terms of power of the tests instead?
- Line 26: the CDFs need to be "centered" around zero, it is not immediately clear what this means (presumably the CDFs need not be symmetric around zero, although perhaps this is the case?). I suggest that this assumption be defined explicitly.
- Line 36: I believe when the authors write "distribution independence" they mean "distribution-free", but "independence" is a loaded word in this context. If the authors do not want to repeat "distribution-free", I would suggest using a few more words but avoid saying that the tests are "distribution independent".
- Line 39: "for somewhat larger computational costs", this phrase sounds a bit awkward, I'd use "at the cost of a larger computational cost", or something along those lines.
- Line 57: "the denominator is a robust estimator for the within-sample variability", I believe this is not accurate, the denominator is an estimator of the squared root of the asymptotic variance of the difference of the estimators (or the estimator of the difference in location parameters), which in this case is a multiple of the within-sample dispersion, but this is just an artifact of using these particular estimators. Also note that I would avoid using words derived from "variance", and rely on "dispersion" instead.
- Lines 69 and 73: efficiency vs power, these terms seem to be used exchangeably here, but although more efficient estimators may result in more powerful tests, the concepts are different. I'd suggest referring to the power of the tests throughout the paper.
- Lines 80/1: same comment as above regarding the reference to the "within-sample variance".
- Lines 108/9: the packages robustbase and RobStatTM also contain functions for tests for the coefficients of linear regression models
- Line 119: "distribution independence", see my comment above.
- The reference to Maronna et al (2006) should be updated to the 2nd edition of the book (2018): [https://www.wiley.com/en-ca/Robust+Statistics:+Theory+and+Methods+(with+R),+2nd+Edition-p-9781119214687](https://www.wiley.com/en-ca/Robust+Statistics:+Theory+and+Methods+(with+R),+2nd+Edition-p-9781119214687)
