### 2022-12-09 #4947 @s-abbas
Hello @msalibian,

Thank you again for your review. We decided to revise the paper, when we receive the suggestions and corrections from the second review. Thus, I only address the first three remarks.

> The LICENSE file is missing.

The LICENSE.md file is in our develop branch. Our plan was to merge the develop branch to the main branch after the review process is finished so that we can submit a new, reviewed version of the package to CRAN. Would this be all right or should we merge the branches now?

> I did not find Community guidelines (for reporting issues or contributing to the project) in the documentation (they are there on the GitHub repo). This is not typically present in R packages stored on CRAN, so I'm inclined to not require it here.

Does this mean that we do not have to change anything or should we present the Community guidelines at some other place?

> Automated tests to verify functionality, I'm happy with the "seeded" examples in one vignette, even if these are not proper unit tests, they do provide a way to check the behaviour of the software.

We implemented several unit tests in the folder "tests". They can be performed, e.g. by calling devtools::test() directly in R from the package folder.

Best regards,
Sermad
