### 2022-12-09 #4947 @msalibian
Hi @s-abbas,
- My guess is that as long as the license file eventually is put in the main branch, then it's OK, but of course I'll defer to @jbytecode on this;
- Again, I think @jbytecode will know better than me what is the exact journal policy about where Community Guidelines should appear;
- Thanks for setting up the tests.

I look forward to reading the revised manuscript.
