### 2022-12-09 #4947 @jbytecode
@msalibian - Yes, during the review process, a suitable license should be provided. And this is current for the Community Guidelines. Thank you for your comments
