### 2022-12-11 #4947 @mingzehuang
Hi, @s-abbas,
I just read your paper and it seems great:) You have clear explanation and comparison with alternative methods:)
I see you have 'Other packages with robust two-sample tests' part at the end of paper. Do you think it might be a part of extension of 'Statement of need'? Probably you may put it right after 'Statement of need'.
Also, do you mind putting a simple small example with code block somewhere in 'Implemented two-sample tests' or 'Applications' to help readers quickly understand the usage of your package since it is a software paper?
