### 2022-12-11 s-abbas/robnptests#1 @mingzehuang
Community Guideline
Hi, @s-abbas,
here is a template to create [community guildline](https://github.com/mingzehuang/latentcor/blob/master/CONTRIBUTING.md) which include how to contribute to the software:)
