+++ paper/REFERENCES.bib
@@ -208,9 +208,13 @@ doi = {10.1007/978-3-030-25147-5_16},
 }
 
 @book{MarMarYoh06robu,
-  title = {Robust statistics: Theory and methods},
-  publisher = {{Wiley, Chichester}},
-  author = {Maronna, R. A. and Martin, D. R. and Yohai, V. J.},
-  year = {2006},
-  doi = {10.1002/0470010940}
+  title = {Robust {{Statistics}}: {{Theory}} and {{Methods}} (with {{R}})},
+  author = {Maronna, Ricardo A. and Martin, Douglas R. and Yohai, Victor J. and Salibián-Barrera, Matías},
+  date = {2019},
+  series = {Wiley Series in Probability and Statistics},
+  edition = {Second edition},
+  publisher = {{Wiley}},
+  location = {{Hoboken, NJ}},
+  doi = {10.1002/9781119214656}
 }
+
