+++ inst/REFERENCES.bib
@@ -83,13 +83,19 @@
   doi = {10.1002/0471725250}
 }
 
+
 @book{MarMarYoh06robu,
-  title = {Robust {{Statistics}}: {{Theory}} and {{Methods}}},
-  publisher = {{Wiley, Chichester}},
-  author = {Maronna, R. A. and Martin, D. R. and Yohai, V. J.},
-  year = {2006},
-  doi = {10.1002/0470010940}
+  title = {Robust {{Statistics}}: {{Theory}} and {{Methods}} (with {{R}})},
+  author = {Maronna, Ricardo A. and Martin, Douglas R. and Yohai, Victor J. and Salibián-Barrera, Matías},
+  date = {2019},
+  series = {Wiley Series in Probability and Statistics},
+  edition = {Second edition},
+  publisher = {{Wiley}},
+  location = {{Hoboken, NJ}},
+  doi = {10.1002/9781119214656}
 }
+
+
 
 @article{Fri12onli,
   title = {On the online estimation of piecewise constant volatilities},
