+++ man/trimmed_test.Rd
@@ -105,9 +105,10 @@ in scale. This is achieved by log-transforming the original squared observations
 i.e. \code{x} is replaced by \code{log(x^2)} and \code{y} by \code{log(y^2)}.
 A potential scale difference then appears as a location difference between
 the transformed samples, see \insertCite{Fri12onli;textual}{robnptests}.
-The sample should not contain zeros to prevent problems with the necessary
-log-transformation. If it contains zeros, uniform noise is added to all
-variables in order to remove zeros and a message is printed.
+Note, that the samples need to be centred around zero. The sample should not
+contain zeros to prevent problems with the necessary log-transformation. If
+it contains zeros, uniform noise is added to all variables in order to remove
+zeros and a message is printed.
 
 If the sample has been modified because of zeros when \code{var.test = TRUE},
 the modified samples can be retrieved using
