+++ paper/REFERENCES.bib
@@ -207,10 +207,10 @@ doi = {10.1007/978-3-030-25147-5_16},
   url = {https://CRAN.R-project.org/package=nptest},
 }
 
-@book{MarMarYoh06robu,
+@book{MarMarYoh19robu,
   title = {Robust {{Statistics}}: {{Theory}} and {{Methods}} (with {{R}})},
   author = {Maronna, Ricardo A. and Martin, Douglas R. and Yohai, Victor J. and Salibián-Barrera, Matías},
-  date = {2019},
+  year = {2019},
   series = {Wiley Series in Probability and Statistics},
   edition = {Second edition},
   publisher = {{Wiley}},
