+++ inst/REFERENCES.bib
@@ -84,10 +84,10 @@
 }
 
 
-@book{MarMarYoh06robu,
+@book{MarMarYoh19robu,
   title = {Robust {{Statistics}}: {{Theory}} and {{Methods}} (with {{R}})},
   author = {Maronna, Ricardo A. and Martin, Douglas R. and Yohai, Victor J. and Salibián-Barrera, Matías},
-  date = {2019},
+  year = {2019},
   series = {Wiley Series in Probability and Statistics},
   edition = {Second edition},
   publisher = {{Wiley}},
