+++ vignettes/robnptests.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Getting started with `robnptests`"
-date: "2022-07-22"
+date: "2022-12-22"
 author: Sermad Abbas, Barbara Brune, Roland Fried
 output:
   rmarkdown::html_document:
@@ -202,7 +202,7 @@ An M-estimator $\hat{\mu}^{(M)}_X$ for the location parameter of the $x$-sample 
   \hat{\mu}^{(M)}_X = \arg\underset{\mu_X}{\min} \sum_{i = 1}^m \rho\left(\frac{x_i - \mu_X}{\theta_X}\right),
 \end{equation*}
 
-where $\rho:\ \mathbb{R} \to \mathbb{R}$ is a function chosen to achieve a "nearly optimal" location estimate when the $X_i$ are normally or approximately normally distributed [@MarMarYoh06robu, p. 22].
+where $\rho:\ \mathbb{R} \to \mathbb{R}$ is a function chosen to achieve a "nearly optimal" location estimate when the $X_i$ are normally or approximately normally distributed [@MarMarYoh19robu, p. 22].
 For differentiable functions $\rho$ with $\rho' = \psi$, the optimization problem can be translated to finding the value of $\mu_X$, for which
 
 \begin{equation*}
@@ -567,7 +567,7 @@ Here is the information about the R-session used for creating this vignette.
 
 ```r
 sessionInfo()
-#> R version 4.2.0 (2022-04-22)
+#> R version 4.2.2 Patched (2022-11-10 r83330)
 #> Platform: x86_64-pc-linux-gnu (64-bit)
 #> Running under: Linux Mint 19.1
 #> 
@@ -576,9 +576,8 @@ sessionInfo()
 #> LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.7.1
 #> 
 #> locale:
-#>  [1] LC_CTYPE=de_DE.UTF-8       LC_NUMERIC=C               LC_TIME=de_DE.UTF-8        LC_COLLATE=de_DE.UTF-8     LC_MONETARY=de_DE.UTF-8   
-#>  [6] LC_MESSAGES=de_DE.UTF-8    LC_PAPER=de_DE.UTF-8       LC_NAME=C                  LC_ADDRESS=C               LC_TELEPHONE=C            
-#> [11] LC_MEASUREMENT=de_DE.UTF-8 LC_IDENTIFICATION=C       
+#>  [1] LC_CTYPE=de_DE.UTF-8       LC_NUMERIC=C               LC_TIME=de_DE.UTF-8        LC_COLLATE=de_DE.UTF-8     LC_MONETARY=de_DE.UTF-8    LC_MESSAGES=de_DE.UTF-8   
+#>  [7] LC_PAPER=de_DE.UTF-8       LC_NAME=C                  LC_ADDRESS=C               LC_TELEPHONE=C             LC_MEASUREMENT=de_DE.UTF-8 LC_IDENTIFICATION=C       
 #> 
 #> attached base packages:
 #> [1] stats     graphics  grDevices utils     datasets  methods   base     
@@ -587,14 +586,13 @@ sessionInfo()
 #> [1] robnptests_1.0.0 testthat_3.1.4  
 #> 
 #> loaded via a namespace (and not attached):
-#>  [1] Rcpp_1.0.9        DEoptimR_1.0-11   compiler_4.2.0    later_1.3.0       urlchecker_1.0.1  prettyunits_1.1.1 profvis_0.3.7     remotes_2.4.2    
[…]
