+++ man/m_est.Rd
@@ -43,7 +43,7 @@ for different psi functions.
 }
 \details{
 To compute the M-estimate, the iterative algorithm described in
-\insertCite{MarMarYoh06robu;textual}{robnptests} is used.
+\insertCite{MarMarYoh19robu;textual}{robnptests} is used.
 The variance is estimated as in \insertCite{Hub81robu;textual}{robnptests}.
 
 If \code{max.it} contains decimal places, it is truncated to an integer
@@ -60,7 +60,7 @@ m_est(x, psi = "huber")
 
 }
 \references{
-\insertRef{MarMarYoh06robu}{robnptests}
+\insertRef{MarMarYoh19robu}{robnptests}
 
 \insertRef{Hub81robu}{robnptests}
 }
