+++ paper/REFERENCES.bib
@@ -142,7 +142,7 @@ doi = {10.1007/978-3-030-25147-5_16},
 }
 
 @Manual{LovMai22walr,
-    title = {walrus: Robust statistical methods},
+    title = {{walrus}: Robust statistical methods},
     author = {Jonathon Love and Patrick Mair},
     year = {2022},
     note = {R package version 1.0.5},
@@ -150,7 +150,7 @@ doi = {10.1007/978-3-030-25147-5_16},
 }
 
 @Manual{Mar20robe,
-    title = {robeth: R functions for robust statistics},
+    title = {{robeth}: R functions for robust statistics},
     author = {Alfio Marazzi},
     year = {2020},
     note = {R package version 2.7-6},
@@ -200,7 +200,7 @@ doi = {10.1007/978-3-030-25147-5_16},
 }
 
 @Manual{Hel21npte,
-  title = {nptest: Nonparametric bootstrap and permutation tests},
+  title = {{nptest}: Nonparametric bootstrap and permutation tests},
   author = {Nathaniel E. Helwig},
   year = {2021},
   note = {R package version 1.0-3},
@@ -218,3 +218,18 @@ doi = {10.1007/978-3-030-25147-5_16},
   doi = {10.1002/9781119214656}
 }
 
+@Manual{YohMarMar22robs,
+  title = {RobStatTM: Robust Statistics: Theory and Methods},
+  author = {Yohai, Victor and Maronna, Ricardo and Martin, Doug and Brownson, Gregory and Konis, Kjell and Salibián-Barrera, Matías},
+  year = {2022},
+  note = {R package version 1.0.5},
+  url = {https://CRAN.R-project.org/package=RobStatTM},
+}
+
+@Manual{MaeRouCro22robu,
+  title = {{robustbase}: Basic robust statistics},
+  author = {Maechler, Martin and Rousseeuw, Peter and Croux, Christophe and Todorov, Valentin and Ruckstuhl, Andreas and Salibián-Barrera, Matías and Verbeke, Tobias and Koller, Manuel and Conceicao, Eduardo L. T. and {di Palma}, Maria {Anna}},
+  year = {2022},
+  note = {R package version 0.95-0},
+  url = {http://CRAN.R-project.org/package=robustbase},
+}
