+++ inst/REFERENCES.bib
@@ -95,8 +95,6 @@
   doi = {10.1002/9781119214656}
 }
 
-
-
 @article{Fri12onli,
   title = {On the online estimation of piecewise constant volatilities},
   author = {Fried, R.},
@@ -108,10 +106,10 @@
   doi = {10.1016/j.csda.2011.02.012}
 }
 
-@Manual{MaeRouCro20robu,
-  title = {robustbase: Basic Robust Statistics},
-  author = {Martin Maechler and Peter Rousseeuw and Christophe Croux and Valentin Todorov and Andreas Ruckstuhl and Matias Salibian-Barrera and Tobias Verbeke and Manuel Koller and Eduardo L. T. Conceicao and Maria {Anna di Palma}},
-  year = {2020},
-  note = {R package version 0.93-6},
-  url = {http://robustbase.r-forge.r-project.org/},
+@Manual{MaeRouCro22robu,
+  title = {{robustbase}: Basic robust statistics},
+  author = {Maechler, Martin and Rousseeuw, Peter and Croux, Christophe and Todorov, Valentin and Ruckstuhl, Andreas and Salibián-Barrera, Matías and Verbeke, Tobias and Koller, Manuel and Conceicao, Eduardo L.T. and {di Palma}, Maria {Anna}},
+  year = {2022},
+  note = {R package version 0.95-0},
+  url = {http://CRAN.R-project.org/package=robustbase},
 }
