### 2022-12-22 c8c524e
documentation - Update paper draft
Move section with other packages that contain robust tests after section "Statemend of need" and add references.
