+++ paper/paper.md
@@ -66,7 +66,20 @@ Common parametric and non-parametric tests for scale differences have similar pr
 In addition, some non-parametric tests for the scale problem do not cope well with asymmetry.
 A possible solution, while retaining the robustness, is to apply the robust location tests to transformed observations as proposed by @Fri12onli.
 Such tests can yield good results in terms of power and size under both asymmetry and outlier corruption.
-However, the tests may be less efficient under symmetry than classical procedures like the Mood test. 
+However, the tests may be less efficient under symmetry than classical procedures like the Mood test.
+
+## Other packages with robust two-sample tests
+The CRAN Task View for robust statistical methods currently lists the three packages `WRS2` [@MaiWil20wrs2], `walrus` [@LovMai22walr], and `robeth` [@Mar20robe] that explicitly deal with robust hypothesis tests for the two-sample problem.
+`WRS2` contains a large collection of robust procedures which are presented in the book *Introduction to Robust Estimation and Hypothesis Testing* by [@Wil17intr]. `walrus` provides a different user interface for the functions in `WRS2`.
+Robust tests can also be found in the packages packages `robeth` [@Mar20robe], `robustbase` [@MaeRouCro22robu], and `RobStatTM` [@YohMarMar22robs].
+However, these tests are for linear hypothesis and thus beyond the scope of `robnptests`.
+
+The functions in `WRS2` concentrate on the heteroscedastic setting, whereas our focus lies on the homoscedastic case.
+The reason of this focus is that especially for small samples, estimating the within-sample variance separately for both samples, as is the case under heteroscedasticity, may lead to unreliable estimates.
+Moreover, choosing equal sample sizes $m = n$, can protect against a deteriorating test performance of our implemented tests under heteroscedasticity in terms of size and power [@StaShe90robu, p. 179].
+
+The R package `nptest` [@Hel21npte] contains nonparametric versions of the two-sample $t$-test, which is realized by using the permutation and randomization principles described above on the $t$-statistic.
+This principle has also been studied in @AbbFri17cont, and, while achieving distribution independence, the test statistic lacks robustness against outliers.
 
 # Implemented two-sample tests
 Each test statistic is a simple ratio inspired by the test statistic of the two-sample $t$-test. The numerator is a robust estimator for the location difference between the two populations and the denominator is a robust estimator for the within-sample variability. 
@@ -101,17 +114,5 @@ Besides conventional two-sample problems, the tests can be applied in a moving t
 @AbbFriGat16dete describe how intensity changes in image sequences generated by a virus sensor are automatically detected by applying the tests to the individual pixel time series of the sequence.
 Moreover, the test statistics can be used as control statistics for robust, (approximately) distribution-free control charts for time series with a time-varying signal [@AbbFri20robu; @AbbFri17cont].
 In @AbbFriHei19dete, the tests are applied to detect unusual sequences in time series of crack widths in concrete beams by searching for sudden scale changes.
-
-## Other packages with robust two-sample tests
-The CRAN Task View for robust statistical methods currently lists the three packages `WRS2` [@MaiWil20wrs2], `walrus` [@LovMai22walr], and `robeth` [@Mar20robe] that explicitly deal with robust hypothesis tests for the two-sample problem.
-`WRS2` contains a large collection of robust procedures which are presented in the book *Introduction to Robust Estimation and Hypothesis Testing* by [@Wil17intr]. `walrus` provides a different user interface for the functions in `WRS2`. 
-The package `robeth` [@Mar20robe] contains some robust tests for linear hypotheses.
-
-The functions in `WRS2` concentrate on the heteroscedastic setting, whereas our focus lies on the homoscedastic case.
[…]
