+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "robnptests -- An R package for robust two-sample location and variability tests"
+title: "robnptests -- An R package for robust two-sample location and dispersion tests"
 tags:
   - R
   - robust statistics
@@ -29,7 +29,7 @@ journal: JOSS
 
 # Summary
 The `robnptests` R package [@R22lang] is a compilation of two-sample tests selected by two criteria: The tests are (i) robust against outliers and (ii) (approximately) distribution free. 
-Regarding the latter aspect, we implemented tests that keep an intended significance level and provide a reasonably high efficiency, both  under a variety of continuous distributions.
+Regarding the latter aspect, we implemented tests that keep an intended significance level and provide a reasonably high power, both  under a variety of continuous distributions.
 Robustness is achieved by using test statistics that are based on robust location and scale measures. 
 
 In what follows, we give a brief description of the package's contents.
@@ -52,9 +52,9 @@ However, non-normality can result in a loss of power [@Wil03appl].
 In addition, the $t$-test is vulnerable to outliers [@FriDeh11robu].
 Distribution-free tests, like the two-sample Wilcoxon rank-sum test, can be nearly as powerful as the $t$-test under normality and may have higher power under non-normality.
 Still, they also can be vulnerable to outliers, particularly for small samples [@FriGat07rank].
-The two-sample tests in `robnptests` combine (approximate) distribution independence and robustness against outliers.
+The two-sample tests in `robnptests` are (approximately) distribution free and, at the same time, robust against outliers.
 Thus, they are well suited for outlier-corrupted samples from unknown data-generating distributions.
-At the same time, such tests can be nearly as powerful as popular procedures like the aforementioned $t$-test or the Wilcoxon test on uncontaminated samples for somewhat larger computational costs.
+At the same time, such tests can be nearly as powerful as popular procedures like the aforementioned $t$-test or the Wilcoxon test on uncontaminated samples for a somewhat longer computation time.
 
 Figure 1 compares the power of the $t$-test, the Wilcoxon test and two robust tests. The HL1-test is based on the one-sample Hodges-Lehmann estimator [@HodLeh63esti] and the Huber M-test uses Huber's M-estimator [@Hub64robu]. We consider a fixed location difference between the samples and a single outlier of increasing size.
 The power of the $t$-test decreases to zero, while the loss in power of the Wilcoxon test and both robust tests is small. 
@@ -66,7 +66,7 @@ Common parametric and non-parametric tests for scale differences have similar pr
 In addition, some non-parametric tests for the scale problem do not cope well with asymmetry.
 A possible solution, while retaining the robustness, is to apply the robust location tests to transformed observations as proposed by @Fri12onli.
 Such tests can yield good results in terms of power and size under both asymmetry and outlier corruption.
-However, the tests may be less efficient under symmetry than classical procedures like the Mood test.
+However, the tests may be less powerful under symmetry than classical procedures like the Mood test.
 
 ## Other packages with robust two-sample tests
 The CRAN Task View for robust statistical methods currently lists the three packages `WRS2` [@MaiWil20wrs2], `walrus` [@LovMai22walr], and `robeth` [@Mar20robe] that explicitly deal with robust hypothesis tests for the two-sample problem.
@@ -75,14 +75,14 @@ Robust tests can also be found in the packages packages `robeth` [@Mar20robe], `
 However, these tests are for linear hypothesis and thus beyond the scope of `robnptests`.
 
 The functions in `WRS2` concentrate on the heteroscedastic setting, whereas our focus lies on the homoscedastic case.
-The reason of this focus is that especially for small samples, estimating the within-sample variance separately for both samples, as is the case under heteroscedasticity, may lead to unreliable estimates.
[…]
