+++ NEWS.md
@@ -1,0 +1,2 @@
+# robnptests 1.0.0
+* Initial CRAN release.
