+++ paper/paper.md
@@ -19,7 +19,7 @@ affiliations:
     name: TU Dortmund University, Faculty of Statistics, 44221 Dortmund, Germany
   - index: 2
     name: Technical University of Vienna, Institute of Statistics and Mathematical Methods in Economics, 1040 Vienna, Austria
-date: 14 October 2022
+date: 12 January 2023
 bibliography: REFERENCES.bib
 output:
   rticles::joss_article
@@ -29,11 +29,11 @@ journal: JOSS
 
 # Summary
 The `robnptests` R package [@R22lang] is a compilation of two-sample tests selected by two criteria: The tests are (i) robust against outliers and (ii) (approximately) distribution free. 
-Regarding the latter aspect, we implemented tests that keep an intended significance level and provide a reasonably high power, both  under a variety of continuous distributions.
+Criterion (ii) means that the implemented tests keep an intended significance level and provide a reasonably high power under a variety of continuous distributions.
 Robustness is achieved by using test statistics that are based on robust location and scale measures. 
 
 In what follows, we give a brief description of the package's contents.
-More details can be found in the introductory vignette of the package, which can be called by `vignette("robnptests")`, and in the cited papers.
+More details can be found in the introductory vignette of the package, which can be opened by calling `vignette("robnptests")` in the R console, and in the cited papers.
 
 # Data situation
 We consider two samples of independent and identically distributed (i.i.d.) random variables $X_1, ..., X_m$ and $Y_1, ..., Y_n$, respectively. 
@@ -64,25 +64,21 @@ The robust tests provide a somewhat higher power than the Wilcoxon test and this
 
 Common parametric and non-parametric tests for scale differences have similar problems as described above for the location tests.
 In addition, some non-parametric tests for the scale problem do not cope well with asymmetry.
-A possible solution, while retaining the robustness, is to apply the robust location tests to transformed observations as proposed by @Fri12onli.
-Such tests can yield good results in terms of power and size under both asymmetry and outlier corruption.
-However, the tests may be less powerful under symmetry than classical procedures like the Mood test.
+The package `robnptests` uses the idea of applying the robust location tests to transformed observations as proposed by @Fri12onli.
+Such tests retain the robustness of the underlying location tests and obtain good results in terms of power and size under both asymmetry and outlier corruption.
+However, these tests may be less powerful under symmetry than classical procedures like the Mood test.
 
-## Other packages with robust two-sample tests
-The CRAN Task View for robust statistical methods currently lists the three packages `WRS2` [@MaiWil20wrs2], `walrus` [@LovMai22walr], and `robeth` [@Mar20robe] that explicitly deal with robust hypothesis tests for the two-sample problem.
-`WRS2` contains a large collection of robust procedures which are presented in the book *Introduction to Robust Estimation and Hypothesis Testing* by [@Wil17intr]. `walrus` provides a different user interface for the functions in `WRS2`.
-Robust tests can also be found in the packages packages `robeth` [@Mar20robe], `robustbase` [@MaeRouCro22robu], and `RobStatTM` [@YohMarMar22robs].
-However, these tests are for linear hypothesis and thus beyond the scope of `robnptests`.
+# Other packages with robust two-sample tests
+The package `WRS2` [@MaiWil20wrs2] contains a collection of robust two-sample location tests for the heteroscedastic setting.
+In `robnptests`, we assume homoscedasticity for the location tests.
+This is because estimating the within-sample dispersion for both samples separately may be unreliable when the sample sizes are small.
+In general, equal sample sizes $m = n$ can protect against a deteriorating performance in terms of size and power [@StaShe90robu, p. 179].
 
-The functions in `WRS2` concentrate on the heteroscedastic setting, whereas our focus lies on the homoscedastic case.
[…]
