+++ man-roxygen/scaleTau2.R
@@ -1,4 +1,4 @@
 #' @param ... additional arguments \code{c1} and \code{c2} that can be passed to
 #'            the function \code{scaleTau2()}, which is used internally for
-#'            estimating the within-sample variability, in order to account for
+#'            estimating the within-sample dispersion, in order to account for
 #'            non-normal distributions; see \insertCite{MarZam02robu;textual}{robnptests}.

[…]
