+++ man-roxygen/wobble_seed.R
@@ -1,6 +1,6 @@
 #' @param wobble.seed an integer value used as a seed for the random number
 #'                    generation in case of \code{wobble = TRUE} or when
-#'                    \code{disp.test = TRUE} with one of the vectors \code{x} and
+#'                    \code{scale_test = TRUE} with one of the vectors \code{x} and
 #'                    \code{y} containing zeros. When no seed is specified, it
 #'                    is chosen randomly and printed in a message. The argument
-#'                    is ignored if \code{disp.test = FALSE} and/or \code{wobble = FALSE}.
+#'                    is ignored if \code{scale_test = FALSE} and/or \code{wobble = FALSE}.
