+++ man-roxygen/var_test.R
@@ -1,2 +1,2 @@
-#' @param disp.test a logical value to specify if the samples should be compared
-#'                 for a difference in scale. The default is \code{disp.test = FALSE}.
+#' @param scale_test a logical value to specify if the samples should be compared
+#'                 for a difference in scale. The default is \code{scale_test = FALSE}.

[…]
