### 2023-01-13 fcd937a
Rename 'rob_var' function to 'rob_scale'
