### 2023-01-13 1da63aa
Rename 'var_test'-Template to 'scale_test'
