+++ R/XXX_asymTrimmedTest.R
@@ -347,7 +347,7 @@
 #' #' @template method
 #' #' @template n_rep
 #' #' @template na_rm
-#' #' @template var_test
+#' #' @template scale_test
 #' #'
 #' #'
 #' #' @details
