### 2023-01-14 e68676b
Update vignette with example for centring for scale tests
