+++ vignettes/robnptests.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Getting started with `robnptests`"
-date: "2022-12-22"
+date: "2023-01-14"
 author: Sermad Abbas, Barbara Brune, Roland Fried
 output:
   rmarkdown::html_document:
@@ -63,6 +63,8 @@ By log-transforming the random variables, we obtain
 \end{align*}
 
 With these transformed samples, we are in the setting of the location problem described in the previous section with $\log\left(\theta_X^2\right)$ instead of $\mu_X$ and $\log\left(\theta_Y^2\right)$ instead of $\mu_Y$ [see @Fri12onli].
+
+
 
 ### Tests for the location problem
 In the following, we describe the location tests implemented in the package.
@@ -323,9 +325,11 @@ x <- rnorm(10)
 y <- rnorm(10)
 
 x
-#>  [1] -0.112892122 -0.381819373 -0.091691759  0.036045206  0.002060011  1.254923407  0.589883973  0.302208595  0.905741048  0.054724887
+#>  [1] -0.112892122 -0.381819373 -0.091691759  0.036045206  0.002060011
+#>  [6]  1.254923407  0.589883973  0.302208595  0.905741048  0.054724887
 y
-#>  [1] -0.8130825  1.3807136 -0.4387136 -1.2788213 -0.6058519  0.9234760  0.5656956 -0.0586360  0.1578270 -0.3858708
+#>  [1] -0.8130825  1.3807136 -0.4387136 -1.2788213 -0.6058519  0.9234760
+#>  [7]  0.5656956 -0.0586360  0.1578270 -0.3858708
 ```
 
 ### Permutation test
@@ -507,10 +511,12 @@ The wobbled samples can be retrieved using the seed printed in the output and th
 set.seed(2187)
 wobble(x1, y1, check = FALSE)
 #> $x
-#>  [1] -0.44559409  0.25924234  0.18034432 -0.42507561 -0.07607574  0.68780214  0.17812437 -0.22174378  0.19595295 -0.14095250
+#>  [1] -0.44559409  0.25924234  0.18034432 -0.42507561 -0.07607574  0.68780214
+#>  [7]  0.17812437 -0.22174378  0.19595295 -0.14095250
 #> 
 #> $y
-#>  [1] -0.04372255  2.22220106  1.23878909  0.16253680 -0.13902692  2.43266503  1.86621893  1.41821762  0.71976748  1.36813320
+#>  [1] -0.04372255  2.22220106  1.23878909  0.16253680 -0.13902692  2.43266503
+#>  [7]  1.86621893  1.41821762  0.71976748  1.36813320
 ```
 
 The argument `check` is used to inspect the samples for duplicated values.
@@ -560,6 +566,62 @@ hl1_test(x = x * sqrt(5), y = y, alternative = "two.sided", delta = 5, method = 
 #>        -0.12240        -1.21528
 ```
 
+If $\mu_X \neq 0$ or $\mu_Y \neq 0$, we need to estimate these parameters from the samples. They can then be subtracted from the respective sample to achieve the required centring.
+
+In the following example, the second sample is centred around 5 instead of 0. The test erroneously rejects the null hypothesis.
+
+
+```r
+set.seed(108)
+
+x <- rnorm(30)
+y <- rnorm(30, mean = 5)
+
+## Asymptotic two-sided test with non-centred variables
[…]
