+++ R/XXX_asymTrimmedTest.R
@@ -1,486 +1,0 @@
-#' #' @title Asymmetrically trimmed mean
-#' #'
-#' #' @description
-#' #' \code{asym_trimmed_mean} calculates an asymmetrically trimmed mean of a sample.
-#' #' In contrast to an ordinary trimmed mean, the numbers of observations removed
-#' #' from the lower and the upper end of the sample do not need to be equal.
-#' #'
-#' #' @template x
-#' #' @template type_skewness
-#' #' @template na_rm
-#' #'
-#' #' @details
-#' #' The number of observations trimmed from the lower and the upper end of the sample
-#' #' depends on the skewness.
-#' #'
-#' #' Three skewness-selector statistics suggested by Reed and Stark (1996) have been
-#' #' implemented. The argument \code{type} specifies which one is used.
-#' #'
-#' #' \describe{
-#' #' \item{"Q2"}{This skewness-selector statistic compares the difference between the upper and the lower 5\% of
-#' #'             the observations to the difference between the upper and lower 50\%.}
-#' #' \item{"SK2"}{This skewness-selector statistic compares the difference between the sample minimum and
-#' #'              median to the difference between the median and the maximum.}
-#' #' \item{"SK5"}{This skewness-selector statistic compares the difference between the sample minimum and
-#' #'              mean to the difference between the mean and the maximum.}
-#' #' }
-#' #'
-#' #' @return
-#' #' The asymmetrically trimmed mean.
-#' #'
-#' #' @importFrom Rdpack reprompt
-#' #'
-#' #' @references
-#' #' \insertRef{ReeSta96hing}{robnptests}
-#' #'
-#' #' @examples
-#' #' ## Generate random sample
-#' #' set.seed(108)
-#' #' x <- rnorm(10)
-#' #'
-#' #' ## Compute asymmetrically trimmed means
-#' #' asym_trimmed_mean(x, type = "Q2")
-#' #' asym_trimmed_mean(x, type = "SK2")
-#' #' asym_trimmed_mean(x, type = "SK5")
-#' #'
-#' #' @export
-#'
-#' asym_trimmed_mean <- function(x, type = c("Q2", "SK2", "SK5"), na.rm = FALSE) {
-#'   ## NA handling
-#'   if (!na.rm & any(is.na(x))) {
-#'     return(NA)
-#'   } else if (na.rm & any(is.na(x))) {
-#'     x <- as.vector(stats::na.omit(x))
-#'   }
-#'
-#'   type <- match.arg(type)
-#'
-#'   ## Sample size and sorted sample
-#'   m <- length(x)
-#'   x.sort <- sort(x)
-#'
-#'   ## Determine lower trimming proportion
-#'   if (type == "Q2") {
-#'
-#'     ## Do not trim, if sample size is too small
-#'     if (floor(0.05 * m) == 0) {
-#'       return(mean(x))
-#'     }
-#'
-#'     gamma <- 0.1
-#'
-#'     U.005 <- mean(x.sort[(m - floor(0.05 * m) + 1):m])
-#'     L.005 <- mean(x.sort[1:floor(0.05 * m)])
-#'     U.05 <- mean(x.sort[(m - floor(0.5 * m) + 1):m])
-#'     L.05 <- mean(x.sort[1:floor(0.5 * m)])
-#'
-#'     gamma.lower <- gamma * (U.005 - L.005)/(U.005 - L.005 + U.05 - L.05)
-#'   } else if (type == "SK2") {
-#'     gamma <- 0.1
-#'
-#'     gamma.lower <- gamma * (min(x) - stats::median(x))/(min(x) - max(x))
-#'   } else if (type == "SK5") {
[…]
