+++ tests/testthat/test_hl2_test.R
@@ -18,10 +18,10 @@ testthat::test_that("hl2_test works correctly", {
 
   testthat::expect_snapshot_output(hl2_test(x = x[1:5], y = y[1:5],
                                             method = "permutation", scale = "S1",
-                                            scale_test = TRUE))
+                                            scale.test = TRUE))
   testthat::expect_snapshot_output(hl2_test(x = x[1:5], y = y[1:5],
                                             method = "permutation", scale = "S2",
-                                            scale_test = TRUE))
+                                            scale.test = TRUE))
 
   # Randomization test
   testthat::expect_snapshot_output(hl2_test(x = x[1:10], y = y[1:10],
@@ -34,16 +34,16 @@ testthat::test_that("hl2_test works correctly", {
   testthat::expect_snapshot_output(hl2_test(x = x[1:10], y = y[1:10],
                                             method = "randomization",
                                             n.rep = 10000,  scale = "S1",
-                                            scale_test = TRUE))
+                                            scale.test = TRUE))
   testthat::expect_snapshot_output(hl2_test(x = x[1:10], y = y[1:10],
                                             method = "randomization",
                                             n.rep = 10000,  scale = "S2",
-                                            scale_test = TRUE))
+                                            scale.test = TRUE))
 
   # Asymptotic test
   testthat::expect_snapshot_output(hl2_test(x = x, y = y, method = "asymptotic"))
   testthat::expect_snapshot_output(hl2_test(x = x, y = y, method = "asymptotic",
-                                            scale_test = TRUE))
+                                            scale.test = TRUE))
 
   # Compare value of the test statistic to manually computed value ----
 
@@ -154,7 +154,7 @@ testthat::test_that("hl2_test works correctly", {
 
   # One of the samples contains zeros
   testthat::expect_message(hl2_test(x = x[1:10], y = c(y[1:9], 0),
-                                    method = "asymptotic", scale_test = TRUE))
+                                    method = "asymptotic", scale.test = TRUE))
 
   # Wobbling ----
 
