+++ man-roxygen/scale_test.R
@@ -1,2 +1,2 @@
-#' @param scale_test a logical value to specify if the samples should be compared
-#'                 for a difference in scale. The default is \code{scale_test = FALSE}.
+#' @param scale.test a logical value to specify if the samples should be compared
+#'                 for a difference in scale. The default is \code{scale.test = FALSE}.
