+++ man-roxygen/wobble_seed_mtest.R
@@ -1,6 +1,6 @@
 #' @param wobble.seed an integer value used as a seed for the random number
-#'                    generation in case that \code{scale_test = TRUE}
+#'                    generation in case that \code{scale.test = TRUE}
 #'                    and one of the vectors \code{x} and
 #'                    \code{y} contains zeros. When no seed is specified, it
 #'                    is chosen randomly and printed in a message. The argument
-#'                    is ignored if \code{scale_test = FALSE}.
+#'                    is ignored if \code{scale.test = FALSE}.

[…]
