### 2023-01-14 f584d3a
Update 'NEWS.md' and '.Rbuildignore'
