+++ NEWS.md
@@ -1,2 +1,8 @@
+# robnptests (development version)
+
+* Renamed the `var.test` argument to `scale.test` in all test functions (`hl1_test()`, `hl2_test()`, `med_test()`, `m_test()`).
+* Renamed the `rob_var()` function to `rob_scale()`.
+
 # robnptests 1.0.0
+
 * Initial CRAN release.
