### 2023-01-14 ceccef4
documentation - Update help pages for functions
