### 2023-01-14 9cc11e4
documentation - Rephrasing in vignette 'robnptests'
