+++ vignettes/robnptests.Rmd
@@ -49,7 +49,7 @@ where $\Delta_0 \in \mathbb{R}$ represents the assumed location difference under
 
 ### Differences in scale
 When setting $\mu_X = \mu_Y = 0$, both data-generating distributions have the location parameter zero but the values of the scale parameter might be unequal.
-We address the following hypothesis pairs:
+We address the following hypotheses:
 \begin{align*}
   &H_0^{(=)}\colon\ \frac{\theta^2_X}{\theta^2_Y} = \tilde{\Delta}_0 \quad \text{vs.} \quad H_1^{(\neq)}\colon\ \frac{\theta^2_X}{\theta^2_Y} \neq \tilde{\Delta}_0 \\
   &H_0^{(>)}\colon\ \frac{\theta^2_X}{\theta^2_Y} \leq \tilde{\Delta}_0 \quad \text{vs.} \quad H_1^{(>)}\colon\ \frac{\theta^2_X}{\theta^2_Y} > \tilde{\Delta}_0 \\
@@ -84,7 +84,7 @@ The package contains tests for the outlined situation that have been proposed in
 Those were selected with the following objectives in mind:
 
 * robustness against outliers,
-* (approximate) distribution independence, i.e. keep $\alpha$ under $H_0$ over a large class of continuous distributions,
+* (approximately) distribution-free, i.e. they keep $\alpha$ under $H_0$ over a large class of continuous distributions,
 * and a large power over many distributions.
 
 In this vignette, we describe the test statistics and how they can be applied using the package's functions.
@@ -325,11 +325,9 @@ x <- rnorm(10)
 y <- rnorm(10)
 
 x
-#>  [1] -0.112892122 -0.381819373 -0.091691759  0.036045206  0.002060011
-#>  [6]  1.254923407  0.589883973  0.302208595  0.905741048  0.054724887
+#>  [1] -0.112892122 -0.381819373 -0.091691759  0.036045206  0.002060011  1.254923407  0.589883973  0.302208595  0.905741048  0.054724887
 y
-#>  [1] -0.8130825  1.3807136 -0.4387136 -1.2788213 -0.6058519  0.9234760
-#>  [7]  0.5656956 -0.0586360  0.1578270 -0.3858708
+#>  [1] -0.8130825  1.3807136 -0.4387136 -1.2788213 -0.6058519  0.9234760  0.5656956 -0.0586360  0.1578270 -0.3858708
 ```
 
 ### Permutation test
@@ -511,12 +509,10 @@ The wobbled samples can be retrieved using the seed printed in the output and th
 set.seed(2187)
 wobble(x1, y1, check = FALSE)
 #> $x
-#>  [1] -0.44559409  0.25924234  0.18034432 -0.42507561 -0.07607574  0.68780214
-#>  [7]  0.17812437 -0.22174378  0.19595295 -0.14095250
+#>  [1] -0.44559409  0.25924234  0.18034432 -0.42507561 -0.07607574  0.68780214  0.17812437 -0.22174378  0.19595295 -0.14095250
 #> 
 #> $y
-#>  [1] -0.04372255  2.22220106  1.23878909  0.16253680 -0.13902692  2.43266503
-#>  [7]  1.86621893  1.41821762  0.71976748  1.36813320
+#>  [1] -0.04372255  2.22220106  1.23878909  0.16253680 -0.13902692  2.43266503  1.86621893  1.41821762  0.71976748  1.36813320
 ```
 
 The argument `check` is used to inspect the samples for duplicated values.
@@ -525,7 +521,7 @@ If `check = TRUE` and all values are unique, the wobbling is omitted.
 # Testing for a difference in scale
[…]
