### 2023-01-14 eea81f7
documentation - Update paper draft and remove unused references .bib file
