+++ paper/REFERENCES.bib
@@ -44,16 +44,6 @@ pages = {251--269},
 doi = {10.1007/978-3-030-25147-5_16},
 }
 
-@article{Abo92robu,
-  title = {Some robust two-sample test statistics based on {{M}}-estimators of location},
-  volume = {21},
-  number = {1},
-  journal = {Communications in Statistics - Simulation and Computation},
-  author = {Aboukalam, M. A. F.},
-  year = {1992},
-  pages = {133 - 148}
-}
-
 @article{Fri12onli,
   title = {On the online estimation of piecewise constant volatilities},
   author = {Fried, Roland},
@@ -141,22 +131,6 @@ doi = {10.1007/978-3-030-25147-5_16},
   doi = {10.1016/C2010-0-67044-1}
 }
 
-@Manual{LovMai22walr,
-    title = {{walrus}: Robust statistical methods},
-    author = {Jonathon Love and Patrick Mair},
-    year = {2022},
-    note = {R package version 1.0.5},
-    url = {https://CRAN.R-project.org/package=walrus},
-}
-
-@Manual{Mar20robe,
-    title = {{robeth}: R functions for robust statistics},
-    author = {Alfio Marazzi},
-    year = {2020},
-    note = {R package version 2.7-6},
-    url = {https://CRAN.R-project.org/package=robeth},
-}
-
 @Manual{R22lang,
   title = {R: A language and environment for statistical computing},
   author = {{R Core Team}},
@@ -216,20 +190,4 @@ doi = {10.1007/978-3-030-25147-5_16},
   publisher = {{Wiley}},
   location = {{Hoboken, NJ}},
   doi = {10.1002/9781119214656}
-}
-
-@Manual{YohMarMar22robs,
-  title = {RobStatTM: Robust Statistics: Theory and Methods},
-  author = {Yohai, Victor and Maronna, Ricardo and Martin, Doug and Brownson, Gregory and Konis, Kjell and Salibián-Barrera, Matías},
-  year = {2022},
-  note = {R package version 1.0.5},
-  url = {https://CRAN.R-project.org/package=RobStatTM},
-}
-
-@Manual{MaeRouCro22robu,
-  title = {{robustbase}: Basic robust statistics},
-  author = {Maechler, Martin and Rousseeuw, Peter and Croux, Christophe and Todorov, Valentin and Ruckstuhl, Andreas and Salibián-Barrera, Matías and Verbeke, Tobias and Koller, Manuel and Conceicao, Eduardo L. T. and {di Palma}, Maria {Anna}},
-  year = {2022},
-  note = {R package version 0.95-0},
-  url = {http://CRAN.R-project.org/package=robustbase},
 }
