+++ vignettes/robnptests.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Getting started with `robnptests`"
-date: "2023-01-14"
+date: "2023-01-18"
 author: Sermad Abbas, Barbara Brune, Roland Fried
 output:
   rmarkdown::html_document:
@@ -26,7 +26,7 @@ We assume the $X_i$ and $Y_j$ can be written as
 \end{align*}
 Here, $\mu_X$ and $\mu_Y$ denote the location parameters, and $\theta_X$ and $\theta_Y$ the scale parameters of the first and the second sample.
 The random variables $\varepsilon_{1}, \ldots, \varepsilon_{m+n}$ are i.i.d. random variables.
-Thus, $\boldsymbol{X}_i$ and $\boldsymbol{Y}_j$ are from the same location-scale family.
+Thus, we assume $\boldsymbol{X}_i$ and $\boldsymbol{Y}_j$ are from the same location-scale family.
 
 ### Differences in location
 Assuming $\theta_X = \theta_Y$, the data-generating distributions differ at most in location.
@@ -48,7 +48,7 @@ Using the tests implemented in `robnptests`, the following hypotheses can be tes
 where $\Delta_0 \in \mathbb{R}$ represents the assumed location difference under $H_0$.
 
 ### Differences in scale
-When setting $\mu_X = \mu_Y = 0$, both data-generating distributions have the location parameter zero but the values of the scale parameter might be unequal.
+When setting $\mu_X = \mu_Y = 0$, both data-generating distributions have location zero, but the values of the scale parameter might be different.
 We address the following hypotheses:
 \begin{align*}
   &H_0^{(=)}\colon\ \frac{\theta^2_X}{\theta^2_Y} = \tilde{\Delta}_0 \quad \text{vs.} \quad H_1^{(\neq)}\colon\ \frac{\theta^2_X}{\theta^2_Y} \neq \tilde{\Delta}_0 \\
@@ -68,7 +68,7 @@ With these transformed samples, we are in the setting of the location problem de
 
 ### Tests for the location problem
 In the following, we describe the location tests implemented in the package.
-We show an example of how use them to test for scale differences at the end of this vignette.
+We show an example of how to use them to test for scale differences at the end of this vignette.
 
 A popular test for the two-sample location problem is the ordinary two-sample $t$-test.
 For small samples, it assumes normality of the data-generating distributions.
@@ -325,9 +325,11 @@ x <- rnorm(10)
 y <- rnorm(10)
 
 x
-#>  [1] -0.112892122 -0.381819373 -0.091691759  0.036045206  0.002060011  1.254923407  0.589883973  0.302208595  0.905741048  0.054724887
+#>  [1] -0.112892122 -0.381819373 -0.091691759  0.036045206  0.002060011
+#>  [6]  1.254923407  0.589883973  0.302208595  0.905741048  0.054724887
 y
-#>  [1] -0.8130825  1.3807136 -0.4387136 -1.2788213 -0.6058519  0.9234760  0.5656956 -0.0586360  0.1578270 -0.3858708
+#>  [1] -0.8130825  1.3807136 -0.4387136 -1.2788213 -0.6058519  0.9234760
+#>  [7]  0.5656956 -0.0586360  0.1578270 -0.3858708
 ```
 
 ### Permutation test
@@ -509,10 +511,12 @@ The wobbled samples can be retrieved using the seed printed in the output and th
 set.seed(2187)
[…]
