### 2023-01-18 723bbd7
Small changes to paper.md
