### 2023-01-19 032cf89
documentation - Rephrasing in vignette for introduction into the package
