+++ vignettes/robnptests.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Getting started with `robnptests`"
-date: "2023-01-18"
+date: "2023-01-19"
 author: Sermad Abbas, Barbara Brune, Roland Fried
 output:
   rmarkdown::html_document:
@@ -126,10 +126,10 @@ Value of argument `alternative` | Alternative hypothesis | Meaning of the altern
 `less`                          | $H_1^{(\geq)}\colon\ \Delta < 0$ | $X$ is stochastically smaller than $Y$
 
 ### MED-tests
-Using the difference of the sample medians, i.e.
+The difference of the sample medians,
 
 \begin{align*}
-   \hat{\Delta}^{(\text{MED})} = \text{median}\left(x_1, \ldots, x_m\right) - \text{median}\left(y_1, \ldots, y_n\right)
+   \hat{\Delta}^{(\text{MED})} = \text{median}\left(x_1, \ldots, x_m\right) - \text{median}\left(y_1, \ldots, y_n\right),
 \end{align*}
 
 is an obvious candidate to estimate the location difference $\Delta$ robustly.
@@ -260,7 +260,7 @@ Let $A$ be the number of theses values, $t_i$ the value of the test statistic fo
 
 and $p$-value $= \frac{A}{B}$.
 
-$B$ increases rapidly in $m$ and $n$, so using the permutation principle can lead to memory and computation time issues. For example, the sample sizes $m = n = 10$ already lead to $\binom{20}{10} = 184\,754$ possible splits.
+$B$ increases rapidly in $m$ and $n$, so using the permutation principle can lead to memory and computation-time issues. For example, the sample sizes $m = n = 10$ already lead to $\binom{20}{10} = 184\,754$ possible splits.
 Thus, we recommend this approach only for small sample sizes or when sufficient computational resources are available.
 
 ### Randomization test
@@ -325,11 +325,9 @@ x <- rnorm(10)
 y <- rnorm(10)
 
 x
-#>  [1] -0.112892122 -0.381819373 -0.091691759  0.036045206  0.002060011
-#>  [6]  1.254923407  0.589883973  0.302208595  0.905741048  0.054724887
+#>  [1] -0.112892122 -0.381819373 -0.091691759  0.036045206  0.002060011  1.254923407  0.589883973  0.302208595  0.905741048  0.054724887
 y
-#>  [1] -0.8130825  1.3807136 -0.4387136 -1.2788213 -0.6058519  0.9234760
-#>  [7]  0.5656956 -0.0586360  0.1578270 -0.3858708
+#>  [1] -0.8130825  1.3807136 -0.4387136 -1.2788213 -0.6058519  0.9234760  0.5656956 -0.0586360  0.1578270 -0.3858708
 ```
 
 ### Permutation test
@@ -511,12 +509,10 @@ The wobbled samples can be retrieved using the seed printed in the output and th
 set.seed(2187)
 wobble(x1, y1, check = FALSE)
 #> $x
-#>  [1] -0.44559409  0.25924234  0.18034432 -0.42507561 -0.07607574  0.68780214
-#>  [7]  0.17812437 -0.22174378  0.19595295 -0.14095250
+#>  [1] -0.44559409  0.25924234  0.18034432 -0.42507561 -0.07607574  0.68780214  0.17812437 -0.22174378  0.19595295 -0.14095250
 #> 
 #> $y
-#>  [1] -0.04372255  2.22220106  1.23878909  0.16253680 -0.13902692  2.43266503
-#>  [7]  1.86621893  1.41821762  0.71976748  1.36813320
[…]
