### 2023-01-19 c04bc58
chore - Update version number in 'DESCRIPTION' and 'NEWS.md'
