+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: robnptests
-Version: 1.0.0
+Version: 1.1.0
 Type: Package
 Title: Robust Nonparametric Two-Sample Tests for Location/Scale
 Authors@R: c(person(given = "Sermad", family = "Abbas", 
