### 2023-01-19 8fc4d39
documentation - Change CRAN URL for vignettes to canonical form
