+++ inst/REFERENCES.bib
@@ -111,5 +111,5 @@
   author = {Maechler, Martin and Rousseeuw, Peter and Croux, Christophe and Todorov, Valentin and Ruckstuhl, Andreas and Salibián-Barrera, Matías and Verbeke, Tobias and Koller, Manuel and Conceicao, Eduardo L.T. and {di Palma}, Maria {Anna}},
   year = {2022},
   note = {R package version 0.95-0},
-  url = {http://CRAN.R-project.org/package=robustbase},
+  url = {https://CRAN.R-project.org/package=robustbase},
 }
