### 2023-01-20 a44aa4c
documentation - Rephrasing in vignette 'robnptests'
