+++ vignettes/robnptests.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Getting started with `robnptests`"
-date: "2023-01-19"
+date: "2023-01-20"
 author: Sermad Abbas, Barbara Brune, Roland Fried
 output:
   rmarkdown::html_document:
@@ -533,7 +533,7 @@ set.seed(108)
 x <- rnorm(30)
 y <- rnorm(30)
 
-## Asymptotic two-sided test for the hypothesis of equal scales for both samples
+## Asymptotic two-sided test for the null hypothesis of equal scales for both samples
 hl1_test(x = x, y = y, alternative = "two.sided", delta = 1, method = "asymptotic", scale.test = TRUE)
 #> 
 #> 	Asymptotic test based on HL1-estimator
@@ -549,7 +549,7 @@ hl1_test(x = x, y = y, alternative = "two.sided", delta = 1, method = "asymptoti
 Including a squared scale ratio of 5 yields
 
 ```r
-## Asymptotic two-sided test for the hypothesis of a scale ratio of 5 between both samples
+## Asymptotic two-sided test for the null hypothesis of a scale ratio of 5 between both samples
 hl1_test(x = x * sqrt(5), y = y, alternative = "two.sided", delta = 5, method = "asymptotic", scale.test = TRUE)
 #> 
 #> 	Asymptotic test based on HL1-estimator
