### 2023-01-20 fe765c1
documentation - Replace URL in 'README.md'
Included full URL to 'CONTRIBUTING.md' because CRAN checks give a NOTE otherwise.
