### 2023-01-21 282276e
documentation - Update help pages for functions
Samples do not necessarily need to be centred around zero for scale tests. It is important that they have equal locations.
