### 2023-01-21 d3170c2
documentation - Rephrasing in vignette 'robnptests'
Changed sentence structure for clarification.
