### 2023-01-21 d2d85af
documentation - Rephrasing in 'paper.md'
