### 2023-01-21 #4947 @s-abbas
Hello @jbytecode, @msalibian, @mingzehuang,

we have finished our revision and pushed a new version to our repository.

We changed the version of the package to 1.1.0 because we plan to release the package on CRAN with this version number after the review. In the first post of this thread, the version is 1.0.0. Is there a way to update the version number?

Regarding the comments of @msalibian:

> The LICENSE file is missing.

We added the LICENSE file in the main and the develop branch.

> I did not find Community guidelines (for reporting issues or contributing to the project) in the documentation (they are there on the GitHub repo). This is not typically present in R packages stored on CRAN, so I'm inclined to not require it here.

The community guidelines can now be found in the file [CONTRIBUTING.md](https://github.com/s-abbas/robnptests/blob/main/CONTRIBUTING.md).

> Title: "location and variability tests", I suggest using "location and dispersion tests". Variability may be perceived as implying the existence of second moments. In the same spirit as the authors use "location", I'd suggest using "dispersion" instead of variability.

We changed the title to "robnptests -- An R package for robust two-sample location and dispersion tests" and replaced the term "variability" by dispersion and scale in the paper and in the package.

> Summary: "high efficiency", efficiency is a concept usually associated with point estimators, not tests. Can this be rephrased in terms of power of the tests instead?

We replaced the term "efficiency" with "power" when speaking of the tests.

> Line 26: the CDFs need to be "centered" around zero, it is not immediately clear what this means (presumably the CDFs need not be symmetric around zero, although perhaps this is the case?). I suggest that this assumption be defined explicitly.

We removed the sentence "Note that F_X and F_Y need to be centred around zero then". The phrasing was quite confusing because we referred to the cumulative distribution function and not the distribution itself. We tried to clarify this by an example in the vignette `vignette("robnptests")` at the end of the section "Testing for a difference in scale". The samples need the same location so that we can safely use the location tests on the log-transformed samples.

> Line 36: I believe when the authors write "distribution independence" they mean "distribution-free", but "independence" is a loaded word in this context. If the authors do not want to repeat "distribution-free", I would suggest using a few more words but avoid saying that the tests are "distribution independent".

We now use the term "distribution free".

> Line 39: "for somewhat larger computational costs", this phrase sounds a bit awkward, I'd use "at the cost of a larger computational cost", or something along those lines.

We rephrased the sentence to
"At the same time, such tests can be nearly as powerful as popular procedures like the aforementioned $t$-test or the Wilcoxon test on uncontaminated samples for a somewhat longer computation time."

> Line 57: "the denominator is a robust estimator for the within-sample variability", I believe this is not accurate, the denominator is an estimator of the squared root of the asymptotic variance of the difference of the estimators (or the estimator of the difference in location parameters), which in this case is a multiple of the within-sample dispersion, but this is just an artifact of using these particular estimators. Also note that I would avoid using words derived from "variance", and rely on "dispersion" instead.

As suggested for title of the paper, we now refer to the dispersion instead of the variability.

> Lines 69 and 73: efficiency vs power, these terms seem to be used exchangeably here, but although more efficient estimators may result in more powerful tests, the concepts are different. I'd suggest referring to the power of the tests throughout the paper.

We now refer to the power of the tests.

> Lines 80/1: same comment as above regarding the reference to the "within-sample variance".

We replaced "variance" by "dispersion".

> Lines 108/9: the packages robustbase and RobStatTM also contain functions for tests for the coefficients of linear regression models

As suggested by @mingzehuang, we moved the section on other packages directly behind "Statement of need". We shortened this section and now refer only to packages that contain two-sample location tests. That is why we removed the other packages mentioned in the previous version of the paper because the applications for which they contain tests are not in the scope of our package. That is why we did not include the packages robustbase and RobStatTM.

> Line 119: "distribution independence", see my comment above.

> The reference to Maronna et al (2006) should be updated to the 2nd edition of the book (2018): https://www.wiley.com/en-ca/Robust+Statistics:+Theory+and+Methods+(with+R),+2nd+Edition-p-9781119214687

We updated the reference.

Regarding the comments of @mingzehuang:

> I see you have 'Other packages with robust two-sample tests' part at the end of paper. Do you think it might be a part of extension of 'Statement of need'? Probably you may put it right after 'Statement of need'.

We moved the section and shortened it.

> Also, do you mind putting a simple small example with code block somewhere in 'Implemented two-sample tests' or 'Applications' to help readers quickly understand the usage of your package since it is a software paper?

We added a small code snippet with an example for a function call. Originally, our idea was to put all code examples into the vignette `vignette("robnptests")` as we can explain them there in more detail. If you do not find the example sufficient or helpful for the readers, we could certainly add more details to in the paper.

Thanks again for the comments and suggestions. :)
