### 2023-02-03 #4947 @jbytecode
Dear @s-abbas

I have just sent a pull request. The PR includes minor text changes in the manuscript. The remaining parts and the bibtex looks fine to me. Please apply the PR if you agree with them.

Thank you in advance.
