### 2023-02-13 8589da2
Updates for robnptests 1.1.0 (#5)
* update manuscript

* documentation - Update manuscript

Add ORCID iD and rephrasing.

* chore - Update DESCRIPTION

Add ORCID iD.

* documentation - Update manuscript

* documentation - Grammar correction in manuscript

---------

Co-authored-by: Mehmet Hakan Satman <mhsatman@gmail.com>
