+++ paper/paper.md
@@ -95,7 +95,7 @@ For scaling, we use different estimators based on medians and pairwise differenc
 
 In addition, we implemented tests based on M-estimators. This approach to robust location estimation allows for flexibility in how outliers are treated through the specification of the tuning constants of the corresponding $\rho$-function. 
 We focus on Huber's $\rho$-function, the bisquare function and Hampel's $\rho$-function.
-The measure for the dispersion within the samples is a pooled statistic derived from the asymptotic normality of the M-estimators [@MarMarYoh19robu, p. 36ff].
+The measure for the dispersion within the samples is a pooled statistic derived from the asymptotic normality of the M-estimators [@MarMarYoh19robu, p. 37ff].
 Moreover, the package contains Yuen's $t$-test which uses the difference of _trimmed means_ to estimate the location difference and a scale estimator based on the pooled winsorized variances [@YueDix73appr].
 
 In case of data with many ties (e.g. caused by discrete sampling), the ties may carry over to the permutation distribution.
