### 2023-02-14 #4947 @s-abbas
Hello @jbytecode,

I have now finished the proof reading of the paper and created the tagged release on GitHub and the Zenodo archive.

>     * [ ]  Please read the paper pdf carefully, one more time, correct any issue if exists. Please perform a full grammar check.

I made some small grammatical changes and corrected a page number in a citation.

>     * [ ]  Create a tagged release in the software repository. Report the version here. The version should be in form of `vx.y.z`, e.g. `v1.2.3`.

The version of the tagged release is [v1.1.0](https://github.com/s-abbas/robnptests/releases/tag/v1.1.0).

>     * [ ]  Create a Zenodo archive of the repository in https://zenodo.org/
>
>     * [ ]  The Zenodo archive (with the same name/title with the JOSS paper) should have the correct metadata, e.g. author names, **ORCID**s, the title of the archive. **Please set the archive title as the paper title.**
>
>     * [ ]  Please provide the tagged release id, zenodo archive DOI, and Zenodo archive URL here.

The Zenodo archive DOI for version v1.1.0 is 10.5281/zenodo.7641636 and the archive URL is https://zenodo.org/record/7641636
